@@ -16,10 +16,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.0.0+1
+version: 1.0.3+4
 
 environment:
-  sdk: ^3.6.0
+  sdk: ^3.7.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -43,6 +43,8 @@ dependencies:
   cached_network_image: ^3.3.0
   equatable: ^2.0.5
   google_sign_in: ^6.2.1
+  crypto: ^3.0.1
+  sign_in_with_apple: ^6.1.1
   pin_code_fields: ^8.0.1
   flutter_rating_bar: ^4.0.1
   table_calendar: ^3.0.9
@@ -58,8 +60,8 @@ dependencies:
   build_runner: ^2.4.8
   json_serializable: ^6.7.1
   bloc: ^8.1.2
-  permission_handler: ^10.4.3
-#  firebase_auth: ^5.1.4
+  permission_handler: ^12.0.0+1
+  firebase_auth: ^5.4.2
   flutter_local_notifications: ^17.2.2
   firebase_messaging: ^15.0.3
   overlay_support: ^2.1.0
@@ -70,12 +72,32 @@ dependencies:
   flutter_rating_stars: ^1.1.0
   dlibphonenumber: ^1.1.25
   lottie: ^3.3.1
-
+  url_launcher: ^6.3.1
   google_maps_flutter: ^2.10.0
   geolocator: ^13.0.2
   geocoding: ^3.0.0
-  cloud_firestore: ^5.6.2
-  firebase_core: ^3.10.1
+  cloud_firestore: ^5.6.3
+  firebase_core: ^3.11.0
+  flutter_zoom_drawer: ^3.2.0
+#  flutter_credit_card: ^4.0.1
+  flutter_multi_formatter: ^2.13.0
+#  u_credit_card: ^1.4.0
+  credit_card_type_detector: ^3.0.0
+  path_provider: ^2.1.5
+  device_info_plus: ^11.3.0
+  flutter_downloader: ^1.12.0
+  file_picker: ^9.0.0
+  open_file: ^3.5.10
+  flutter_widget_from_html: ^0.15.3
+  lucide_icons: ^0.257.0
+  change_app_package_name: ^1.4.0
+  flutter_stripe: ^11.4.0
+  cloud_functions: ^5.3.3
+  flutter_dotenv: ^5.2.1
+  flutter_staggered_grid_view: ^0.7.0
+  firebase_remote_config: ^5.4.2
+  firebase_analytics: ^11.4.4
+    #  firebase_functions: ^4.1.1
     # The following adds the Cupertino Icons font to your application.
     # Use with the CupertinoIcons class for iOS style icons.
     #  path_provider: ^2.1.4
@@ -111,6 +133,8 @@ flutter:
     - assets/
     - assets/icons/
     - assets/images/
+
+
 
 
   # An image asset can refer to one or more resolution-specific "variants", see
