@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:lottie/lottie.dart';
 import 'package:nb_utils/nb_utils.dart';
+import 'package:primax_lyalaty_program/screens/dashboard_screen/dashboard_screen.dart';
+import 'package:primax_lyalaty_program/screens/home_screen/home_screen.dart';
 import 'package:primax_lyalaty_program/screens/login_screen/login_screen.dart';
 import 'package:primax_lyalaty_program/screens/onboard_screen/onboard_screen.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -27,7 +29,7 @@ class _SplashScreenState extends State<SplashScreen> {
 
    Future.delayed(const Duration(seconds: 2), () {
      if (onboardingComplete) {
-       LoginScreen().launch(context,isNewTask: true,pageRouteAnimation: PageRouteAnimation.Fade);
+       DashboardScreen().launch(context,isNewTask: true,pageRouteAnimation: PageRouteAnimation.Fade);
 
        // Navigator.pushReplacementNamed(context, '/home');
      } else {
@@ -46,7 +48,19 @@ class _SplashScreenState extends State<SplashScreen> {
            height: double.maxFinite,
             child: Image.asset('assets/images/img_splash.png',fit: BoxFit.cover,),
           ),
-          Center(child: Image.asset('assets/images/app_logo.png',height: 200,width: 200,)),
+          Center(child: Column(
+            crossAxisAlignment: CrossAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
+            spacing: 10,
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              ClipRRect(
+                  borderRadius: BorderRadius.circular(30), // Adjust radius as needed
+
+                  child: Image.asset('assets/images/app_logo.png',fit: BoxFit.cover,height: 120,width: 120,)),
+              Text('Primax',style: TextStyle(fontSize: 20,fontWeight: FontWeight.w600,),),
+            ],
+          )),
           Positioned(
               bottom: 50,
               left: 0,
